Splitted CheckoutPage into two components
